requirement.txt
astroid==2.4.2
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.9.2
cycler==0.10.0
Django==2.2
django-filter==2.3.0
djangorestframework==3.11.0
Faker==4.1.1
future==0.18.2
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
matplotlib==3.2.2
mccabe==0.6.1
netmiko==3.1.1
numpy==1.19.0
opencv-python==4.2.0.34
paramiko==2.7.1
pycparser==2.20
pylint==2.5.3
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyserial==3.4
python-dateutil==2.8.1
PythonTurtle==0.3.1
pytz==2020.1
scp==0.13.2
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
textfsm==1.1.0
toml==0.10.1
wrapt==1.12.1
